pyproject.toml
[build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "scale"
version = "0.1.0"
description = ""
authors = [ "Enrico <enrico.rotundo@gmail.com>",]
readme = "README.md"
[[tool.poetry.packages]]
include = "scale"
from = "."

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
aiofiles = "^23.2.1"
pydantic = "^2.5.3"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
httpx = "^0.27.0"
backoff = "^2.2.1"
docker = "^7.0.0"
celery = "^5.4.0"
redis = "^5.0.1"
pyjwt = "^2.8.0"
psutil = "^5.9.7"
pytz = "^2024.1"
gitpython = "^3.1.40"
poetry-core = "^1.8.1"
poetry = "^1.7.1"
setuptools = "^69.0.3"
pyyaml = "^6.0.1"
protobuf = "^5.26.1"
ecdsa = "^0.19.0"
ipfshttpclient = "^0.7.0"
grpcio = "^1.66.2"
grpcio-tools = "^1.66.2"
charset-normalizer = "3.3.2"
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
aiohttp = "^3.11.9"
psycopg2 = "^2.9.10"
importlib-metadata = "^8.6.1"
watchdog = "^6.0.0"
flower = "^2.0.1"

